refactor: extract flash out of server.go
