+++ internal/buildapi/server.go
@@ -17,7 +17,6 @@ import (
 	"path"
 	"regexp"
 	"sort"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -992,20 +991,6 @@ func (a *APIServer) handleUploadFiles(c *gin.Context) {
 	a.uploadFiles(c, name)
 }
 
-// setupLogStreamHeaders configures HTTP headers for log streaming
-func setupLogStreamHeaders(c *gin.Context) {
-	c.Writer.Header().Set("Content-Type", "text/plain; charset=utf-8")
-	c.Writer.Header().Set("Transfer-Encoding", "chunked")
-	c.Writer.Header().Set("Cache-Control", "no-cache, no-store, must-revalidate")
-	c.Writer.Header().Set("Connection", "keep-alive")
-	c.Writer.Header().Set("X-Accel-Buffering", "no")
-	c.Writer.Header().Set("X-Content-Type-Options", "nosniff")
-	c.Writer.Header().Set("Pragma", "no-cache")
-	c.Writer.WriteHeader(http.StatusOK)
-	_, _ = c.Writer.Write([]byte("Waiting for logs...\n"))
-	c.Writer.Flush()
-}
-
 // getStepContainerNames returns container names for pipeline steps
 func getStepContainerNames(pod corev1.Pod) []string {
 	stepNames := make([]string, 0, len(pod.Spec.Containers))
@@ -1499,67 +1484,6 @@ func createPushSecret(
 	}
 
 	return secretName, nil
-}
-
-// Shell metacharacters that must be blocked to prevent injection attacks
-var shellMetachars = []string{";", "|", "&", "$", "`", "(", ")", "{", "}", "<", ">", "!", "\\", "'", "\"", "\n", "\r"}
-
-// validateInput validates a string for dangerous characters and length
-func validateInput(value, fieldName string, maxLen int, allowEmpty bool, extraChars ...string) error {
-	if value == "" {
-		if allowEmpty {
-			return nil
-		}
-		return fmt.Errorf("%s is required", fieldName)
-	}
-
-	// Combine shell metacharacters with any additional blocked characters
-	blockedChars := append(shellMetachars, extraChars...)
-	for _, char := range blockedChars {
-		if strings.Contains(value, char) {
-			return fmt.Errorf("%s contains invalid character: %q", fieldName, char)
-		}
-	}
-
-	if len(value) > maxLen {
-		return fmt.Errorf("%s too long (max %d characters)", fieldName, maxLen)
-	}
-	return nil
-}
-
-func validateContainerRef(ref string) error {
-	return validateInput(ref, "container reference", 500, true)
-}
-
-func validateBuildName(name string) error {
-	if err := validateInput(name, "build name", 253, false, "/"); err != nil {
-		return err
-	}
-
-	// Check if name would become empty after sanitization
-	sanitized := sanitizeBuildNameForValidation(name)
-	if sanitized == "" {
-		return fmt.Errorf("build name contains only invalid characters")
-	}
-
-	return nil
-}
-
-func sanitizeBuildNameForValidation(name string) string {
-	name = strings.ToLower(name)
-	var b strings.Builder
-	for _, r := range name {
-		if (r >= 'a' && r <= 'z') || (r >= '0' && r <= '9') || r == '-' {
-			b.WriteRune(r)
-		} else {
-			b.WriteRune('-')
-		}
-	}
-	result := strings.ReplaceAll(b.String(), "--", "-")
-	for strings.Contains(result, "--") {
-		result = strings.ReplaceAll(result, "--", "-")
-	}
-	return strings.Trim(result, "-")
 }
 
 // validateBuildRequest validates the build request, sanitizes the name, and applies defaults
@@ -2882,137 +2806,6 @@ func createFlashClientSecret(
 	return c.Create(ctx, secret)
 }
 
-func writeJSON(c *gin.Context, status int, v any) {
-	c.Header("Cache-Control", "no-store")
-	c.IndentedJSON(status, v)
-}
-
-const maxPageLimit = 500
-
-// parsePagination extracts limit and offset from query parameters.
-// When limit is not provided, 0 is returned and applyPagination returns
-// the full slice (preserving backward compatibility for existing clients).
-// When provided, limit is clamped to maxPageLimit (500).
-func parsePagination(c *gin.Context) (limit, offset int) {
-	if l := c.Query("limit"); l != "" {
-		if n, err := strconv.Atoi(l); err == nil && n > 0 {
-			limit = n
-			if limit > maxPageLimit {
-				limit = maxPageLimit
-			}
-		}
-	}
-	if o := c.Query("offset"); o != "" {
-		if n, err := strconv.Atoi(o); err == nil && n >= 0 {
-			offset = n
-		}
-	}
-	return
-}
-
-// applyPagination returns the paginated window of items. A limit of 0
-// means "no limit" — the full slice (from offset) is returned.
-func applyPagination[T any](items []T, limit, offset int) []T {
-	if offset >= len(items) {
-		return []T{}
-	}
-	if limit <= 0 {
-		return items[offset:]
-	}
-	end := offset + limit
-	if end > len(items) {
-		end = len(items)
-	}
-	return items[offset:end]
-}
-
-func parseSinceTime(sinceParam string) *metav1.Time {
-	if sinceParam == "" {
-		return nil
-	}
-	t, err := time.Parse(time.RFC3339, sinceParam)
-	if err != nil {
-		return nil
-	}
-	return &metav1.Time{Time: t}
-}
-
-func resolveNamespace() string {
-	if ns := strings.TrimSpace(os.Getenv("BUILD_API_NAMESPACE")); ns != "" {
-		return ns
-	}
-	if data, err := os.ReadFile("/var/run/secrets/kubernetes.io/serviceaccount/namespace"); err == nil {
-		ns := strings.TrimSpace(string(data))
-		if ns != "" {
-			return ns
-		}
-	}
-	return "default"
-}
-
-func getRESTConfigFromRequest(_ *gin.Context) (*rest.Config, error) {
-	var cfg *rest.Config
-	var err error
-	cfg, err = rest.InClusterConfig()
-	if err != nil {
-		kubeconfig := os.Getenv("KUBECONFIG")
-		cfg, err = clientcmd.BuildConfigFromFlags("", kubeconfig)
-		if err != nil {
-			return nil, fmt.Errorf("failed to build kube config: %w", err)
-		}
-	}
-	cfgCopy := rest.CopyConfig(cfg)
-	cfgCopy.Timeout = 30 * time.Minute
-	return cfgCopy, nil
-}
-
-// getKubernetesClient creates a controller-runtime client for accessing Kubernetes resources
-func getKubernetesClient() (client.Client, error) {
-	cfg, err := rest.InClusterConfig()
-	if err != nil {
-		kubeconfig := os.Getenv("KUBECONFIG")
-		cfg, err = clientcmd.BuildConfigFromFlags("", kubeconfig)
-		if err != nil {
-			return nil, fmt.Errorf("failed to build kube config: %w", err)
-		}
-	}
-
-	scheme := runtime.NewScheme()
-	if err := automotivev1alpha1.AddToScheme(scheme); err != nil {
-		return nil, fmt.Errorf("failed to add scheme: %w", err)
-	}
-
-	k8sClient, err := client.New(cfg, client.Options{Scheme: scheme})
-	if err != nil {
-		return nil, fmt.Errorf("failed to create k8s client: %w", err)
-	}
-	return k8sClient, nil
-}
-
-func getClientFromRequest(c *gin.Context) (client.Client, error) {
-	cfg, err := getRESTConfigFromRequest(c)
-	if err != nil {
-		return nil, err
-	}
-
-	scheme := runtime.NewScheme()
-	if err := automotivev1alpha1.AddToScheme(scheme); err != nil {
-		return nil, fmt.Errorf("failed to add automotive scheme: %w", err)
-	}
-	if err := corev1.AddToScheme(scheme); err != nil {
-		return nil, fmt.Errorf("failed to add core scheme: %w", err)
-	}
-	if err := tektonv1.AddToScheme(scheme); err != nil {
-		return nil, fmt.Errorf("failed to add tekton scheme: %w", err)
-	}
-
-	k8sClient, err := client.New(cfg, client.Options{Scheme: scheme})
-	if err != nil {
-		return nil, fmt.Errorf("failed to create k8s client: %w", err)
-	}
-	return k8sClient, nil
-}
-
 // refreshAuthConfigIfNeeded periodically checks and refreshes authentication configuration from OperatorConfig
 // IMPORTANT: This function only recreates the OIDC authenticator if the config actually changed.
 func (a *APIServer) refreshAuthConfigIfNeeded() {
@@ -3274,15 +3067,6 @@ func extractBearerToken(c *gin.Context) string {
 	return ""
 }
 
-func (a *APIServer) resolveRequester(c *gin.Context) string {
-	if v, ok := c.Get("requester"); ok {
-		if username, ok := v.(string); ok && username != "" {
-			return username
-		}
-	}
-	return "unknown"
-}
-
 // handleGetAuthConfig returns OIDC configuration for clients (no auth required)
 func (a *APIServer) handleGetAuthConfig(c *gin.Context) {
 	// Refresh auth config if needed
@@ -3442,30 +3226,6 @@ func (a *APIServer) handleGetOperatorConfig(c *gin.Context) {
 	c.JSON(http.StatusOK, response)
 }
 
-// Flash API handlers
-
-func (a *APIServer) handleCreateFlash(c *gin.Context) {
-	a.log.Info("create flash", "reqID", c.GetString("reqID"))
-	a.createFlash(c)
-}
-
-func (a *APIServer) handleListFlash(c *gin.Context) {
-	a.log.Info("list flash jobs", "reqID", c.GetString("reqID"))
-	a.listFlash(c)
-}
-
-func (a *APIServer) handleGetFlash(c *gin.Context) {
-	name := c.Param("name")
-	a.log.Info("get flash", "flash", name, "reqID", c.GetString("reqID"))
-	a.getFlash(c, name)
-}
-
-func (a *APIServer) handleFlashLogs(c *gin.Context) {
-	name := c.Param("name")
-	a.log.Info("flash logs requested", "flash", name, "reqID", c.GetString("reqID"))
-	a.streamFlashLogs(c, name)
-}
-
 // Sealed API handlers
 
 // sealedPathToOperation maps the API path prefix to the AIB sealed operation.
@@ -3508,428 +3268,6 @@ func (a *APIServer) handleSealedLogs(c *gin.Context) {
 	name := c.Param("name")
 	a.log.Info("reseal logs requested", "name", name, "reqID", c.GetString("reqID"))
 	a.streamSealedLogs(c, name)
-}
-
-func (a *APIServer) createFlash(c *gin.Context) {
-	var req FlashRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid JSON request"})
-		return
-	}
-
-	// Validate required fields
-	if req.ImageRef == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "imageRef is required"})
-		return
-	}
-	if req.ClientConfig == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "clientConfig is required"})
-		return
-	}
-
-	// Auto-generate name if not provided
-	if req.Name == "" {
-		req.Name = fmt.Sprintf("flash-%s", uuid.New().String()[:5])
-	}
-
-	// Validate name
-	if err := validateBuildName(req.Name); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	// Validate mutual exclusivity of lease-name and lease-duration
-	if req.LeaseName != "" && req.LeaseDuration != "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "lease-name and lease-duration are mutually exclusive"})
-		return
-	}
-
-	k8sClient, err := getClientFromRequest(c)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("k8s client error: %v", err)})
-		return
-	}
-
-	restCfg, err := getRESTConfigFromRequest(c)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-	clientset, err := kubernetes.NewForConfig(restCfg)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	ctx := c.Request.Context()
-	namespace := resolveNamespace()
-	requestedBy := a.resolveRequester(c)
-
-	// Load OperatorConfig for target mappings, image overrides, and lease duration defaults
-	operatorConfig := &automotivev1alpha1.OperatorConfig{}
-	if err := k8sClient.Get(ctx, types.NamespacedName{Name: "config", Namespace: namespace}, operatorConfig); err != nil {
-		if !k8serrors.IsNotFound(err) {
-			a.log.Error(err, "failed to load OperatorConfig for flash, using defaults")
-		}
-		operatorConfig = &automotivev1alpha1.OperatorConfig{}
-	}
-
-	// Resolve exporter selector and flash command from OperatorConfig
-	exporterSelector, flashCmd := resolveFlashTargetConfig(req, operatorConfig)
-	if exporterSelector == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "exporterSelector or valid target is required"})
-		return
-	}
-
-	// Replace placeholders in flash command
-	if flashCmd != "" {
-		flashCmd = strings.ReplaceAll(flashCmd, "{image_uri}", req.ImageRef)
-		flashCmd = strings.ReplaceAll(flashCmd, "{artifact_url}", req.ImageRef)
-	}
-
-	// Create Jumpstarter client config secret
-	secretName, createdSecret, secretErr := createFlashClientConfigSecret(ctx, clientset, namespace, req)
-	if secretErr != nil {
-		c.JSON(secretErr.code, gin.H{"error": secretErr.message})
-		return
-	}
-
-	// Create OCI auth secret for flash image pull credentials
-	flashOCIAuthSecretName, createdOCIAuthSecret, ociErr := createFlashOCIAuthSecret(ctx, clientset, namespace, req.Name, req.RegistryCredentials)
-	if ociErr != nil {
-		_ = clientset.CoreV1().Secrets(namespace).Delete(ctx, secretName, metav1.DeleteOptions{})
-		c.JSON(ociErr.code, gin.H{"error": ociErr.message})
-		return
-	}
-
-	// Build task config from OperatorConfig for flash task generation
-	var flashBuildConfig *tasks.BuildConfig
-	if operatorConfig.Spec.OSBuilds != nil {
-		flashBuildConfig = &tasks.BuildConfig{
-			FlashTimeoutMinutes:  operatorConfig.Spec.OSBuilds.GetFlashTimeoutMinutes(),
-			DefaultLeaseDuration: operatorConfig.Spec.Jumpstarter.GetDefaultLeaseDuration(),
-		}
-	}
-
-	// Get the flash task spec
-	flashTask := tasks.GenerateFlashTask(namespace, flashBuildConfig)
-
-	// Lease duration: only resolve when not using an existing lease
-	// Fallback: request > FlashTimeoutMinutes (as HH:MM:SS) > Jumpstarter default > constant
-	leaseDuration := req.LeaseDuration
-	if req.LeaseName == "" && leaseDuration == "" {
-		if operatorConfig.Spec.OSBuilds != nil && operatorConfig.Spec.OSBuilds.FlashTimeoutMinutes > 0 {
-			m := operatorConfig.Spec.OSBuilds.FlashTimeoutMinutes
-			leaseDuration = fmt.Sprintf("%02d:%02d:00", m/60, m%60)
-		} else {
-			leaseDuration = operatorConfig.Spec.Jumpstarter.GetDefaultLeaseDuration()
-		}
-	}
-
-	// Build workspace bindings
-	workspaces := []tektonv1.WorkspaceBinding{
-		{
-			Name: "jumpstarter-client",
-			Secret: &corev1.SecretVolumeSource{
-				SecretName: secretName,
-			},
-		},
-	}
-	if flashOCIAuthSecretName != "" {
-		workspaces = append(workspaces, tektonv1.WorkspaceBinding{
-			Name: "flash-oci-auth",
-			Secret: &corev1.SecretVolumeSource{
-				SecretName: flashOCIAuthSecretName,
-			},
-		})
-	}
-
-	// Create the flash TaskRun
-	taskRun := &tektonv1.TaskRun{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      req.Name,
-			Namespace: namespace,
-			Labels: map[string]string{
-				"app.kubernetes.io/managed-by": "build-api",
-				"app.kubernetes.io/part-of":    "automotive-dev",
-				"app.kubernetes.io/name":       "flash-taskrun",
-				flashTaskRunLabel:              req.Name,
-			},
-			Annotations: map[string]string{
-				"automotive.sdv.cloud.redhat.com/requested-by": requestedBy,
-				"automotive.sdv.cloud.redhat.com/image-ref":    req.ImageRef,
-			},
-		},
-		Spec: tektonv1.TaskRunSpec{
-			ServiceAccountName: automotivev1alpha1.BuildServiceAccountName,
-			TaskSpec:           &flashTask.Spec,
-			Params: []tektonv1.Param{
-				{Name: "image-ref", Value: tektonv1.ParamValue{Type: tektonv1.ParamTypeString, StringVal: req.ImageRef}},
-				{Name: "exporter-selector", Value: tektonv1.ParamValue{Type: tektonv1.ParamTypeString, StringVal: exporterSelector}},
-				{Name: "flash-cmd", Value: tektonv1.ParamValue{Type: tektonv1.ParamTypeString, StringVal: flashCmd}},
-				{Name: "lease-duration", Value: tektonv1.ParamValue{Type: tektonv1.ParamTypeString, StringVal: leaseDuration}},
-				{Name: "lease-name", Value: tektonv1.ParamValue{Type: tektonv1.ParamTypeString, StringVal: req.LeaseName}},
-			},
-			Workspaces: workspaces,
-		},
-	}
-
-	if err := k8sClient.Create(ctx, taskRun); err != nil {
-		// Clean up secrets if TaskRun creation fails
-		_ = clientset.CoreV1().Secrets(namespace).Delete(ctx, secretName, metav1.DeleteOptions{})
-		if flashOCIAuthSecretName != "" {
-			_ = clientset.CoreV1().Secrets(namespace).Delete(ctx, flashOCIAuthSecretName, metav1.DeleteOptions{})
-		}
-		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("failed to create flash TaskRun: %v", err)})
-		return
-	}
-
-	// Set owner reference on secrets for automatic cleanup
-	ownerRef := []metav1.OwnerReference{
-		{
-			APIVersion: "tekton.dev/v1",
-			Kind:       "TaskRun",
-			Name:       taskRun.Name,
-			UID:        taskRun.UID,
-		},
-	}
-	createdSecret.OwnerReferences = ownerRef
-	if _, err := clientset.CoreV1().Secrets(namespace).Update(ctx, createdSecret, metav1.UpdateOptions{}); err != nil {
-		log.Printf("WARNING: failed to set owner reference on secret %s: %v", secretName, err)
-	}
-	if createdOCIAuthSecret != nil {
-		createdOCIAuthSecret.OwnerReferences = ownerRef
-		if _, updErr := clientset.CoreV1().Secrets(namespace).Update(ctx, createdOCIAuthSecret, metav1.UpdateOptions{}); updErr != nil {
-			log.Printf("WARNING: failed to set owner reference on flash OCI auth secret %s: %v", flashOCIAuthSecretName, updErr)
-		}
-	}
-
-	writeJSON(c, http.StatusAccepted, FlashResponse{
-		Name:        req.Name,
-		Phase:       phasePending,
-		Message:     "Flash TaskRun created",
-		RequestedBy: requestedBy,
-		TaskRunName: taskRun.Name,
-	})
-}
-
-func (a *APIServer) listFlash(c *gin.Context) {
-	namespace := resolveNamespace()
-	limit, offset := parsePagination(c)
-
-	k8sClient, err := getClientFromRequest(c)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("k8s client error: %v", err)})
-		return
-	}
-
-	ctx := c.Request.Context()
-
-	// List TaskRuns with flash label
-	taskRunList := &tektonv1.TaskRunList{}
-	if err := k8sClient.List(ctx, taskRunList, client.InNamespace(namespace), client.HasLabels{flashTaskRunLabel}); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("failed to list flash TaskRuns: %v", err)})
-		return
-	}
-
-	// Sort by creation time, newest first
-	sort.Slice(taskRunList.Items, func(i, j int) bool {
-		return taskRunList.Items[j].CreationTimestamp.Before(&taskRunList.Items[i].CreationTimestamp)
-	})
-
-	page := applyPagination(taskRunList.Items, limit, offset)
-
-	resp := make([]FlashListItem, 0, len(page))
-	for _, tr := range page {
-		phase, message := getTaskRunStatus(&tr)
-		var compStr string
-		if tr.Status.CompletionTime != nil {
-			compStr = tr.Status.CompletionTime.Format(time.RFC3339)
-		}
-		resp = append(resp, FlashListItem{
-			Name:           tr.Name,
-			Phase:          phase,
-			Message:        message,
-			RequestedBy:    tr.Annotations["automotive.sdv.cloud.redhat.com/requested-by"],
-			CreatedAt:      tr.CreationTimestamp.Format(time.RFC3339),
-			CompletionTime: compStr,
-		})
-	}
-	writeJSON(c, http.StatusOK, resp)
-}
-
-func (a *APIServer) getFlash(c *gin.Context, name string) {
-	namespace := resolveNamespace()
-
-	k8sClient, err := getClientFromRequest(c)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("k8s client error: %v", err)})
-		return
-	}
-
-	ctx := c.Request.Context()
-	taskRun := &tektonv1.TaskRun{}
-	if err := k8sClient.Get(ctx, types.NamespacedName{Name: name, Namespace: namespace}, taskRun); err != nil {
-		if k8serrors.IsNotFound(err) {
-			c.JSON(http.StatusNotFound, gin.H{"error": "flash TaskRun not found"})
-			return
-		}
-		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("failed to get flash TaskRun: %v", err)})
-		return
-	}
-
-	// Verify it's a flash TaskRun
-	if taskRun.Labels[flashTaskRunLabel] == "" {
-		c.JSON(http.StatusNotFound, gin.H{"error": "flash TaskRun not found"})
-		return
-	}
-
-	phase, message := getTaskRunStatus(taskRun)
-	var startStr, compStr string
-	if taskRun.Status.StartTime != nil {
-		startStr = taskRun.Status.StartTime.Format(time.RFC3339)
-	}
-	if taskRun.Status.CompletionTime != nil {
-		compStr = taskRun.Status.CompletionTime.Format(time.RFC3339)
-	}
-
-	writeJSON(c, http.StatusOK, FlashResponse{
-		Name:           taskRun.Name,
-		Phase:          phase,
-		Message:        message,
-		RequestedBy:    taskRun.Annotations["automotive.sdv.cloud.redhat.com/requested-by"],
-		StartTime:      startStr,
-		CompletionTime: compStr,
-		TaskRunName:    taskRun.Name,
-	})
-}
-
-func getTaskRunStatus(tr *tektonv1.TaskRun) (phase, message string) {
-	// Check if completed
-	if tr.Status.CompletionTime != nil {
-		// Check conditions for success/failure
-		for _, cond := range tr.Status.Conditions {
-			if cond.Type == "Succeeded" {
-				if cond.Status == corev1.ConditionTrue {
-					return phaseCompleted, "Flash completed successfully"
-				}
-				return phaseFailed, cond.Message
-			}
-		}
-		return phaseFailed, "Flash failed"
-	}
-
-	// Check if running
-	if tr.Status.StartTime != nil {
-		return phaseRunning, "Flash in progress"
-	}
-
-	return phasePending, "Waiting to start"
-}
-
-func (a *APIServer) streamFlashLogs(c *gin.Context, name string) {
-	namespace := resolveNamespace()
-
-	k8sClient, err := getClientFromRequest(c)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("k8s client error: %v", err)})
-		return
-	}
-
-	restCfg, err := getRESTConfigFromRequest(c)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-	clientset, err := kubernetes.NewForConfig(restCfg)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	ctx := c.Request.Context()
-
-	// Verify the TaskRun exists and is a flash TaskRun
-	taskRun := &tektonv1.TaskRun{}
-	if err := k8sClient.Get(ctx, types.NamespacedName{Name: name, Namespace: namespace}, taskRun); err != nil {
-		if k8serrors.IsNotFound(err) {
-			c.JSON(http.StatusNotFound, gin.H{"error": "flash TaskRun not found"})
-			return
-		}
-		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("failed to get flash TaskRun: %v", err)})
-		return
-	}
-	if taskRun.Labels[flashTaskRunLabel] == "" {
-		c.JSON(http.StatusNotFound, gin.H{"error": "flash TaskRun not found"})
-		return
-	}
-
-	sinceTime := parseSinceTime(c.Query("since"))
-	streamDuration := time.Duration(a.limits.MaxLogStreamDurationMinutes) * time.Minute
-	streamCtx, cancel := context.WithTimeout(ctx, streamDuration)
-	defer cancel()
-
-	// Get the pod name from TaskRun status
-	podName := taskRun.Status.PodName
-	if podName == "" {
-		c.JSON(http.StatusServiceUnavailable, gin.H{"error": "flash pod not ready"})
-		return
-	}
-
-	setupLogStreamHeaders(c)
-
-	// TaskRun pods use step containers with naming convention "step-<step-name>"
-	containerName := "step-flash"
-
-	// Stream logs
-	req := clientset.CoreV1().Pods(namespace).GetLogs(podName, &corev1.PodLogOptions{
-		Container: containerName,
-		Follow:    true,
-		SinceTime: sinceTime,
-	})
-	stream, err := req.Stream(streamCtx)
-	if err != nil {
-		_, _ = fmt.Fprintf(c.Writer, "\n[Error streaming logs: %v]\n", err)
-		c.Writer.Flush()
-		return
-	}
-	defer func() {
-		if err := stream.Close(); err != nil {
-			fmt.Fprintf(os.Stderr, "Warning: failed to close stream: %v\n", err)
-		}
-	}()
-
-	_, _ = c.Writer.Write([]byte("\n===== Flash TaskRun Logs =====\n\n"))
-	c.Writer.Flush()
-
-	scanner := bufio.NewScanner(stream)
-	scanner.Buffer(make([]byte, 64*1024), 1024*1024)
-
-	for scanner.Scan() {
-		select {
-		case <-streamCtx.Done():
-			return
-		default:
-		}
-		line := scanner.Bytes()
-		if _, writeErr := c.Writer.Write(line); writeErr != nil {
-			return
-		}
-		if _, writeErr := c.Writer.Write([]byte("\n")); writeErr != nil {
-			return
-		}
-		c.Writer.Flush()
-	}
-
-	if err := scanner.Err(); err != nil && err != io.EOF {
-		var errMsg []byte
-		errMsg = fmt.Appendf(errMsg, "\n[Stream error: %v]\n", err)
-		_, _ = c.Writer.Write(errMsg)
-		c.Writer.Flush()
-	}
-
-	_, _ = c.Writer.Write([]byte("\n[Log streaming completed]\n"))
-	c.Writer.Flush()
 }
 
 // validateSealedRequest validates and normalizes a SealedRequest, returning the resolved stages or an error message.
